update calMonthlyInterest method to return double; adjust related tests and output values

+++ src/main/java/domain/amount/FixedDepositAmount.java
@@ -29,7 +29,7 @@ public class FixedDepositAmount implements LumpSumInvestmentAmount {
 	}
 
 	@Override
-	public int calMonthlyInterest(InterestRate interestRate) {
+	public double calMonthlyInterest(InterestRate interestRate) {
 		return interestRate.calMonthlyInterest(amount);
 	}
 }
